backend/requirements.txt
APScheduler==3.11.1
asgiref==3.10.0
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
cffi==2.0.0
channels==4.3.1
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
packaging==25.0
psutil==7.1.3
psycopg2-binary==2.9.9
requests==2.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==25.3.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzlocal==5.3.1
websockets==12.0
zope.interface==8.1.1
